third_party/WebKit/LayoutTests/external/wpt/webmidi/midi-disabled-by-feature-policy.https.sub.html

<!DOCTYPE html>
<body>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=resources/featurepolicytest.js></script>
<script>
'use strict';

promise_test(() => {
  return getFeatureTestPromise().then(() => {
    assert_unreached('expected promise to reject with ' + getFeatureError());
  }, error => {
    assert_equals(error.name, getFeatureError());
  });
}, 'Feature-Policy header {"' + getFeatureName() + '": []} disallows the top-level document.');

async_test(t => {
  assert_available_in_iframe(t, '', '#' + getFeatureError());
}, 'Feature-Policy header {"' + getFeatureName() + '": []} disallows same-origin iframes.');

async_test(t => {
  assert_available_in_iframe(t, 'https://{{domains[www]}}:{{ports[https][0]}}', '#' + getFeatureError());
}, 'Feature-Policy header {"' + getFeatureName() + '": []} disallows cross-origin iframes.');
</script>
</body>
